For Gabebe
I moved most of the stuff to the character class and gave them their
proper names, right now there's no time limiter or animation
continuation.  It seems so slowly slide the player for as long as they
key press and it sort of queues the process and gives a sliding effect
if a key is held down too long.  Probably need to mix in the clock too
for timing whether or not a keypress after walking is active.